requirements.txt
# Core Data Processing
pyspark==3.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# GCP
google-cloud-bigquery==3.17.1
google-cloud-storage==2.14.0

# Data Quality
great-expectations==0.18.8

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Utilities
numpy==1.24.4
pandas==2.1.4
pyarrow==14.0.2
langdetect==1.0.9
beautifulsoup4==4.12.3
requests==2.31.0

# Development & Testing
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.15
mypy==1.8.0
pre-commit==3.6.0

# Orchestration
apache-airflow==2.8.1
